+++ apps/netconfapi/src/main/java/org/onosproject/netconfapi/CallhomeProtocolProxy.java
@@ -1,0 +1,19 @@
+package org.onosproject.netconfapi;
+
+
+import org.onlab.rest.AbstractWebApplication;
+
+import java.util.Set;
+
+/**
+ * NetconfAPI Server front-end application.
+ */
+public class CallhomeProtocolProxy extends AbstractWebApplication {
+
+    @Override
+    public Set<Class<?>> getClasses() {
+        return getClasses(
+                CallhomeWebResources.class
+        );
+    }
+}
