+++ apps/netconfapi/src/main/java/org/onosproject/netconfapi/package-info.java
@@ -1,0 +1,1 @@
+package org.onosproject.netconfapi;
