+++ apps/netconfapi/src/main/resources/definitions/Callhome.json
