+++ web/api/src/main/java/org/onosproject/rest/resources/CoreWebApplication.java
@@ -56,8 +56,7 @@ public class CoreWebApplication extends AbstractWebApplication {
                 SystemInfoWebResource.class,
                 PacketProcessorsWebResource.class,
                 AuditFilter.class,
-                ModulationWebResource.class,
-                CallhomeWebResources.class
+                ModulationWebResource.class
         );
     }
 }
